Remove separate "Run" target.
