+++ Package.swift
@@ -14,7 +14,7 @@ let package = Package(
         .package(url: "https://github.com/vapor/leaf.git", from: "4.0.0"),{{/leaf}}
     ],
     targets: [
-        .target(
+        .executableTarget(
             name: "App",
             dependencies: [{{#fluent}}
                 .product(name: "Fluent", package: "fluent"),
@@ -29,7 +29,6 @@ let package = Package(
                 .unsafeFlags(["-cross-module-optimization"], .when(configuration: .release))
             ]
         ),
-        .executableTarget(name: "Run", dependencies: [.target(name: "App")]),
         .testTarget(name: "AppTests", dependencies: [
             .target(name: "App"),
             .product(name: "XCTVapor", package: "vapor"),
