+++ Sources/App/entrypoint.swift
@@ -1,0 +1,15 @@
+import Vapor
+
+@main
+enum Entrypoint {
+    static func main() throws {
+        var env = try Environment.detect()
+        try LoggingSystem.bootstrap(from: &env)
+        
+        let app = Application(env)
+        defer { app.shutdown() }
+        
+        try configure(app)
+        try app.run()
+    }
+}
